use-cases/dataset-finder-integration/Dataset Finder/website/requirements.txt
fuzzywuzzy==0.18.0
flask==2.3.2
python-Levenshtein==0.21.1
flask_httpauth==4.8.0
flask_session==0.5.0
matplotlib==3.8.0
schedule==1.2.1
beautifulsoup4==4.12.3
requests-oauthlib==1.1.0
PyJWT==2.8.0
Flask-Mail==0.9.1